+++ allauth/socialaccount/apps.py
@@ -1,6 +1,5 @@
 from django.apps import AppConfig
-
-from allauth.compat import ugettext_lazy as _
+from django.utils.translation import gettext_lazy as _
 
 
 class SocialAccountConfig(AppConfig):