+++ allauth/socialaccount/models.py
@@ -6,15 +6,12 @@ from django.contrib.sites.shortcuts import get_current_site
 from django.core.exceptions import PermissionDenied
 from django.db import models
 from django.utils.crypto import get_random_string
+from django.utils.encoding import force_str
+from django.utils.translation import gettext_lazy as _
 
 import allauth.app_settings
 from allauth.account.models import EmailAddress
 from allauth.account.utils import get_next_redirect_url, setup_user_email
-from allauth.compat import (
-    force_str,
-    python_2_unicode_compatible,
-    ugettext_lazy as _,
-)
 from allauth.utils import get_user_model
 
 from ..utils import get_request_param
@@ -39,7 +36,6 @@ class SocialAppManager(models.Manager):
         return app
 
 
-@python_2_unicode_compatible
 class SocialApp(models.Model):
     objects = SocialAppManager()
 
@@ -73,7 +69,6 @@ class SocialApp(models.Model):
         return self.name
 
 
-@python_2_unicode_compatible
 class SocialAccount(models.Model):
     user = models.ForeignKey(allauth.app_settings.USER_MODEL,
                              on_delete=models.CASCADE)
@@ -128,7 +123,6 @@ class SocialAccount(models.Model):
         return self.get_provider().wrap_account(self)
 
 
-@python_2_unicode_compatible
 class SocialToken(models.Model):
     app = models.ForeignKey(SocialApp, on_delete=models.CASCADE)
     account = models.ForeignKey(SocialAccount, on_delete=models.CASCADE)