+++ allauth/socialaccount/providers/authentiq/views.py
@@ -1,6 +1,7 @@
 import requests
 
-from allauth.compat import urljoin
+from urllib.parse import urljoin
+
 from allauth.socialaccount import app_settings
 from allauth.socialaccount.providers.oauth2.views import (
     OAuth2Adapter,