+++ ChangeLog.rst
@@ -1,3 +1,10 @@
+...
+
+Backwards incompatible changes
+------------------------------
+- Dropped Python 2 and Django 1 compatibility
+
+
 0.40.0 (2019-08-29)
 *******************